refactor: move packages from stoppropaganda to internal
