+++ internal/sockshttp/socks4.go
