+++ internal/spdnsclient/hook.go
