+++ internal/spdnsclient/net.go
