+++ internal/spdnsclient/unixlike_dnsclient.go
