+++ internal/spdnsclient/unixlike_dnsconfig.go
